+++ src/App.tsx
@@ -2,6 +2,7 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router'
 import { PlayerProvider } from '@/contexts/PlayerContext'
 import { HomePage } from '@/pages/HomePage.tsx'
 import { AlbumDetailPage } from '@/pages/AlbumDetailPage'
+import { LibraryPage } from '@/pages/LibraryPage'
 
 function App() {
     return (
@@ -10,6 +11,7 @@ function App() {
                 <Routes>
                     <Route path="/" element={<HomePage />} />
                     <Route path="/album/:albumId" element={<AlbumDetailPage />} />
+                    <Route path={"/library"} element={<LibraryPage />} />
                 </Routes>
             </Router>
         </PlayerProvider>
