+++ src/pages/content/AlbumDetailContent.tsx
@@ -1,0 +1,290 @@
+// components/oktaf/AlbumDetailContent.tsx
+import { Button } from '@/components/ui/button.tsx'
+import { AlbumArt } from '@/components/ui/AlbumArt.tsx'
+import { usePlayer } from '@/contexts/PlayerContext.tsx'
+import { useAlbumColors } from '@/hooks/useAlbumColors.ts'
+import {
+    navigationIcons,
+    getAlbumTracks,
+    calculateAlbumDuration,
+    formatPlayCount,
+    type Album,
+    type CurrentTrack,
+    type Track
+} from '@/data/DummyData.tsx'
+
+const {
+    Play,
+    Heart,
+    MoreHorizontal,
+    Shuffle,
+    ThumbsDown,
+    Download
+} = navigationIcons
+
+type AlbumDetailContentProps = {
+    album: Album
+    albumId: string
+}
+
+export function AlbumDetailContent({ album, albumId }: AlbumDetailContentProps) {
+    const {
+        playTrack,
+        playAlbum,
+        currentTrack,
+        isPlaying,
+        playerState,
+        toggleShuffle,
+        currentAlbum
+    } = usePlayer()
+
+    const albumArtUrl = typeof album?.art === 'string' ? album.art : album?.art?.value || ''
+    const { dominant, vibrant, muted, isLoading: colorsLoading } = useAlbumColors(albumArtUrl)
+
+    const tracks = getAlbumTracks(album.id)
+    const albumDuration = tracks.length > 0 ? calculateAlbumDuration(tracks) : '0m'
+    const isCurrentAlbum = currentAlbum === albumId
+
+    const handleTrackPlay = (track: Track) => {
+        const trackData: CurrentTrack = {
+            id: track.id,
+            title: track.title,
+            artist: track.artist,
+            album: album?.title || '',
+            art: album?.art || {
+                type: 'gradient',
+                value: 'bg-gradient-to-r from-pink-500 via-red-500 to-yellow-500',
+                alt: 'Default gradient background'
+            },
+            duration: track.duration,
+            currentTime: '0:00',
+            isLiked: track.isLiked || false,
+            isDisliked: track.isDisliked || false,
+            isBookmarked: track.isBookmarked || false,
+        }
+        playTrack(trackData, albumId)
+    }
+
+    const handlePlayAlbum = () => {
+        if (!albumId || tracks.length === 0) return
+        playAlbum(albumId)
+    }
+
+    const handleShufflePlay = () => {
+        if (!albumId || tracks.length === 0) return
+        if (!playerState.isShuffled) {
+            toggleShuffle()
+        }
+        setTimeout(() => {
+            playAlbum(albumId)
+        }, 50)
+    }
+
+    return (
+        <>
+            {/* Header with album info and dynamic background */}
+            <div className="relative min-h-[500px]">
+                {/* Background with album art blur effect */}
+                <div
+                    className="absolute inset-0 bg-cover bg-center bg-no-repeat"
+                    style={{
+                        backgroundImage: `url(${albumArtUrl})`,
+                        filter: 'blur(50px) brightness(0.3)',
+                        transform: 'scale(1.1)',
+                    }}
+                />
+
+                {/* Dynamic color overlays */}
+                {colorsLoading ? (
+                    <>
+                        <div className="absolute inset-0 bg-gradient-to-b from-black/20 via-black/40 to-black/90" />
+                        <div className="absolute inset-0 bg-gradient-to-r from-black/60 via-transparent to-black/60" />
+                    </>
+                ) : (
+                    <>
+                        <div
+                            className="absolute inset-0"
+                            style={{
+                                background: `linear-gradient(135deg, ${dominant}75 0%, ${vibrant}65 50%, ${muted}80 100%)`
+                            }}
+                        />
+                        <div
+                            className="absolute inset-0"
+                            style={{
+                                background: `radial-gradient(ellipse 80% 60% at center top, ${vibrant}55 0%, transparent 70%)`
+                            }}
+                        />
+                        <div className="absolute inset-0 bg-gradient-to-r from-black/40 via-transparent to-black/40" />
+                        <div className="absolute inset-0 bg-gradient-to-b from-transparent via-black/25 to-black/85" />
+                    </>
+                )}
+
+                {/* Header content with top padding for floating nav */}
+                <div className="relative z-20 p-8 pt-24">
+                    {/* Album info section */}
+                    <div className="flex items-end gap-8 mb-8">
+                        <AlbumArt
+                            art={album.art}
+                            size="xl"
+                            className="w-56 h-56 shadow-2xl rounded-lg flex-shrink-0"
+                        />
+
+                        <div className="flex-1 pb-4">
+                            <div className="text-sm text-white/70 mb-2 font-medium">Album</div>
+                            <h1 className="text-6xl font-bold mb-6 leading-tight text-white drop-shadow-lg">
+                                {album.title}
+                            </h1>
+                            <div className="flex items-center gap-2 text-white/90">
+                                <span className="font-semibold text-white">{album.artist}</span>
+                                <span className="text-white/60">•</span>
+                                <span className="text-white/80">{album.year}</span>
+                                <span className="text-white/60">•</span>
+                                <span className="text-white/80">{tracks.length} songs</span>
+                                <span className="text-white/60">•</span>
+                                <span className="text-white/80">{albumDuration}</span>
+                            </div>
+                        </div>
+                    </div>
+
+                    {/* Action buttons */}
+                    <div className="flex items-center gap-6">
+                        <Button
+                            size="icon"
+                            className="bg-green-500 hover:bg-green-400 text-black w-14 h-14 rounded-full shadow-lg transition-all duration-200 hover:scale-105"
+                            onClick={handlePlayAlbum}
+                            title="Play album"
+                        >
+                            <Play className="w-6 h-6 fill-current ml-1" />
+                        </Button>
+
+                        <Button
+                            variant="ghost"
+                            size="icon"
+                            onClick={handleShufflePlay}
+                            className={`w-10 h-10 transition-all duration-200 hover:scale-110 ${
+                                isCurrentAlbum && playerState.isShuffled
+                                    ? 'text-green-500 hover:text-green-400'
+                                    : 'text-white/70 hover:text-white hover:bg-white/10'
+                            }`}
+                            title="Shuffle play"
+                        >
+                            <Shuffle className="w-5 h-5" />
+                        </Button>
+
+                        <Button
+                            variant="ghost"
+                            size="icon"
+                            className="text-white/70 hover:text-white hover:bg-white/10 w-10 h-10 transition-all duration-200 hover:scale-110"
+                            title="Like album"
+                        >
+                            <Heart className="w-5 h-5" />
+                        </Button>
+
+                        <Button
+                            variant="ghost"
+                            size="icon"
+                            className="text-white/70 hover:text-white hover:bg-white/10 w-10 h-10 transition-all duration-200 hover:scale-110"
+                            title="Dislike album"
+                        >
+                            <ThumbsDown className="w-5 h-5" />
+                        </Button>
+
+                        <Button
+                            variant="ghost"
+                            size="icon"
+                            className="text-white/70 hover:text-white hover:bg-white/10 w-10 h-10 transition-all duration-200 hover:scale-110"
+                            title="Download album"
+                        >
+                            <Download className="w-5 h-5" />
+                        </Button>
+
+                        <Button
+                            variant="ghost"
+                            size="icon"
+                            className="text-white/70 hover:text-white hover:bg-white/10 w-10 h-10 transition-all duration-200 hover:scale-110"
+                            title="More options"
+                        >
+                            <MoreHorizontal className="w-5 h-5" />
+                        </Button>
+                    </div>
+                </div>
+            </div>
+
+            {/* Track listing with dynamic background continuation */}
+            <div
+                className="px-8 pb-8"
+                style={{
+                    background: colorsLoading
+                        ? 'linear-gradient(to bottom, rgba(0,0,0,0.9), #0a0a0a)'
+                        : `linear-gradient(to bottom, ${dominant}15, #0a0a0a)`
+                }}
+            >
+                {tracks.length > 0 ? (
+                    <>
+                        {/* Table header */}
+                        <div className="grid grid-cols-12 gap-4 py-4 px-4 text-white/100 text-sm font-medium border-b border-white/10 mb-2">
+                            <div className="col-span-1 text-center">#</div>
+                            <div className="col-span-3">Title</div>
+                            <div className="col-span-3 text-right">Plays</div>
+                            <div className="col-span-3 text-right">Duration</div>
+                        </div>
+
+                        {/* Track rows */}
+                        <div className="space-y-1">
+                            {tracks.map((track, index) => {
+                                const isCurrentTrack = currentTrack?.id === track.id
+                                const isCurrentlyPlaying = isCurrentTrack && isPlaying
+
+                                return (
+                                    <div
+                                        key={track.id}
+                                        className={`grid grid-cols-12 gap-4 py-3 px-4 rounded-lg transition-all duration-200 group cursor-pointer ${
+                                            isCurrentTrack
+                                                ? 'bg-white/10 text-green-400'
+                                                : 'hover:bg-white/5 text-white'
+                                        }`}
+                                        onClick={() => handleTrackPlay(track)}
+                                    >
+                                        <div className="col-span-1 text-center text-sm flex items-center justify-center">
+                                            {isCurrentlyPlaying ? (
+                                                <div className="flex gap-1 items-center">
+                                                    <div className="w-1 h-3 bg-green-400 rounded-full animate-pulse"></div>
+                                                    <div className="w-1 h-2 bg-green-400 rounded-full animate-pulse delay-75"></div>
+                                                    <div className="w-1 h-4 bg-green-400 rounded-full animate-pulse delay-150"></div>
+                                                </div>
+                                            ) : (
+                                                <>
+                                                    <span className={`group-hover:hidden ${isCurrentTrack ? 'text-green-400' : 'text-white/60'}`}>
+                                                        {track.number || index + 1}
+                                                    </span>
+                                                    <Play className="w-4 h-4 hidden group-hover:inline-block fill-current text-white" />
+                                                </>
+                                            )}
+                                        </div>
+                                        <div className="col-span-3">
+                                            <div className={`text-sm font-medium ${isCurrentTrack ? 'text-green-400' : 'text-white'}`}>
+                                                {track.title}
+                                            </div>
+                                            <div className="text-white/60 text-xs mt-1">{track.artist}</div>
+                                        </div>
+                                        <div className="col-span-3 text-right text-white/60 text-sm flex items-center justify-end">
+                                            {formatPlayCount(track.plays)}
+                                        </div>
+                                        <div className="col-span-3 text-right text-white/60 text-sm flex items-center justify-end">
+                                            {track.duration}
+                                        </div>
+                                    </div>
+                                )
+                            })}
+                        </div>
+                    </>
+                ) : (
+                    <div className="text-center py-16">
+                        <div className="text-white/60 text-xl mb-2">No tracks available for this album</div>
+                        <div className="text-white/40 text-sm">Track data will be added soon</div>
+                    </div>
+                )}
+            </div>
+        </>
+    )
+}
